+++ src/core/providers/strategies/LedgerProviderStrategy/LedgerProviderStrategy.ts
@@ -2,11 +2,10 @@ import { Message, Transaction } from '@multiversx/sdk-core/out';
 import { IDAppProviderOptions } from '@multiversx/sdk-dapp-utils/out';
 import { HWProvider } from '@multiversx/sdk-hw-provider';
 import { safeWindow } from 'constants/index';
-import {
-  LedgerConnectStateManager,
-  PendingTransactionsEventsEnum,
-  PendingTransactionsStateManager
-} from 'core/managers';
+
+import { LedgerConnectStateManager } from 'core/managers/internal/LedgerConnectStateManager/LedgerConnectStateManager';
+import { PendingTransactionsStateManager } from 'core/managers/internal/PendingTransactionsStateManager/PendingTransactionsStateManager';
+import { PendingTransactionsEventsEnum } from 'core/managers/internal/PendingTransactionsStateManager/types/pendingTransactions.types';
 import { getAddress } from 'core/methods/account/getAddress';
 import { getIsLoggedIn } from 'core/methods/account/getIsLoggedIn';
 import { IProvider } from 'core/providers/types/providerFactory.types';
@@ -16,6 +15,7 @@ import {
   LedgerConnectModal,
   PendingTransactionsModal
 } from 'lib/sdkDappCoreUi';
+import { SigningWarningsEnum } from 'types/enums.types';
 import { ProviderErrorsEnum } from 'types/provider.types';
 import { createUIElement } from 'utils/createUIElement';
 import { getLedgerProvider } from './helpers';
@@ -48,13 +48,13 @@ export class LedgerProviderStrategy {
     this.address = address || '';
   }
 
-  public createProvider = async (options: {
+  public createProvider = async (options?: {
     anchor?: HTMLElement;
   }): Promise<IProvider> => {
     this.initialize();
     await defineCustomElements(safeWindow);
 
-    const eventBus = await this.createEventBus(options.anchor);
+    const eventBus = await this.createEventBus(options?.anchor);
 
     if (eventBus) {
       this.manager = new LedgerConnectStateManager(eventBus);
@@ -178,34 +178,44 @@ export class LedgerProviderStrategy {
     };
   };
 
-  private signMessage = async (message: Message) => {
-    if (!this.provider || !this._signMessage) {
-      throw new Error(ProviderErrorsEnum.notInitialized);
-    }
+  private signMessage = async (message: Message): Promise<Message> => {
+    const msg = await new Promise<Awaited<Message>>(async (resolve, reject) => {
+      if (!this.provider || !this._signMessage) {
+        return reject(ProviderErrorsEnum.notInitialized);
+      }
 
-    const modalElement = await createUIElement<PendingTransactionsModal>({
-      name: 'pending-transactions-modal'
+      const modalElement = await createUIElement<PendingTransactionsModal>({
+        name: 'pending-transactions-modal'
+      });
+
+      const { eventBus, manager, onClose } =
+        await this.getModalHandlers(modalElement);
+
+      const closeModal = () => {
+        onClose();
+        reject({ message: SigningWarningsEnum.cancelled });
+      };
+
+      eventBus.subscribe(PendingTransactionsEventsEnum.CLOSE, closeModal);
+
+      manager.updateData({
+        isPending: true,
+        title: 'Message Signing',
+        subtitle: 'Check your Ledger device to sign the message'
+      });
+
+      try {
+        const signedMessage = await this._signMessage(message);
+
+        resolve(signedMessage);
+      } catch (err) {
+        reject(err);
+      } finally {
+        onClose();
+        eventBus.unsubscribe(PendingTransactionsEventsEnum.CLOSE, closeModal);
+      }
     });
-
-    const { eventBus, manager, onClose } =
-      await this.getModalHandlers(modalElement);
-
-    eventBus.subscribe(PendingTransactionsEventsEnum.CLOSE, onClose);
-
-    manager.updateData({
-      isPending: true,
-      title: 'Message Signing',
-      subtitle: 'Check your Ledger device to sign the message'
-    });
-
-    try {
-      const signedMessage = await this._signMessage(message);
-
-      return signedMessage;
-    } finally {
-      onClose();
-      eventBus.unsubscribe(PendingTransactionsEventsEnum.CLOSE, onClose);
-    }
+    return msg;
   };
 
   private getModalHandlers = async (modalElement: PendingTransactionsModal) => {