+++ src/core/providers/strategies/helpers/signTransactions/signTransactions.ts
@@ -6,13 +6,14 @@ import {
   GAS_PRICE_MODIFIER,
   MULTI_TRANSFER_EGLD_TOKEN
 } from 'constants/mvx.constants';
-import { SignTransactionsStateManager } from 'core/managers/SignTransactionsStateManager/SignTransactionsStateManager';
+import { SignTransactionsStateManager } from 'core/managers/internal/SignTransactionsStateManager/SignTransactionsStateManager';
 import {
   ISignTransactionsModalData,
   SignEventsEnum
-} from 'core/managers/SignTransactionsStateManager/types';
+} from 'core/managers/internal/SignTransactionsStateManager/types';
 import { getAddress } from 'core/methods/account/getAddress';
 import { getEgldLabel } from 'core/methods/network/getEgldLabel';
+import { cancelCrossWindowAction } from 'core/providers/helpers/cancelCrossWindowAction';
 import { IProvider } from 'core/providers/types/providerFactory.types';
 import { SignTransactionsModal } from 'lib/sdkDappCoreUi';
 import { formatAmount } from 'lib/sdkDappUtils';
@@ -190,8 +191,9 @@ export async function signTransactions({
         manager.updateData(data);
       };
 
-      const onCancel = () => {
+      const onCancel = async () => {
         reject(new Error('Transaction signing cancelled by user'));
+        await cancelCrossWindowAction();
         signModalElement.remove();
       };