+++ src/lib/sdkWebWalletIframeProvider.ts
@@ -1,0 +1,1 @@
+export { IframeProvider } from '@multiversx/sdk-web-wallet-iframe-provider/out';